Edit configuration of application.
