+++ src/server.ts
@@ -1,0 +1,6 @@
+import { PORT } from './common/config';
+import app from './app';
+
+app.listen(PORT, (): void =>
+  console.log(`App is running on http://localhost:${PORT}`)
+);
